backend/pyproject.toml
[tool.poetry]
name = "rag-chatbot-backend"
version = "0.1.0"
description = "Backend for RAG Chatbot Integration with Physical AI & Humanoid Robotics Book"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
qdrant-client = "^1.7.0"
openai = "^1.3.5"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.23"
python-dotenv = "^1.0.0"
asyncpg = "^0.29.0"
pyjwt = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
aiofiles = "^23.2.1"
requests = "^2.31.0"
tiktoken = "^0.5.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "scripts.dev:main"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
skip-string-normalization = true

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88